allow linking solitary posts
needs tests

+++ src/router/helpers.rs
@@ -47,7 +47,7 @@ pub fn dev() -> bool {
     std::env::var("DEV").is_ok_and(|v| v == "1")
 }
 
-pub fn per_page() -> i32 {
+pub fn per_page() -> usize {
     match std::env::var("PER_PAGE") {
         Ok(per_page) => per_page.parse().expect("parse PER_PAGE env"),
         Err(_) => 1_000,
